update balance column type

+++ database/connection.py
@@ -47,7 +47,7 @@ def init_db():
             id SERIAL PRIMARY KEY,
             mint VARCHAR(255) NOT NULL UNIQUE,
             symbol VARCHAR(50),
-            balance INTEGER NOT NULL,
+            balance NUMERIC NOT NULL,
             buy_price DECIMAL(20, 8),
             usdt_value DECIMAL(20, 8) NOT NULL
         )
